+++ src/main/java/KzeroJun/livestream/auth/dto/request/SignUpRequest.java
@@ -1,0 +1,27 @@
+package KzeroJun.livestream.auth.dto.request;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Pattern;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class SignUpRequest {
+
+	@NotBlank
+	@Pattern(regexp = "^[0-9A-Za-z]([-_.]?[0-9a-zA-Z])*@[0-9a-zA-Z]([-_.]?[0-9a-zA-Z])+[.][a-zA-Z]{2,3}$")
+	private String email;
+
+	@NotBlank
+	private String password;
+
+	@NotBlank
+	private String nickname;
+
+}
