+++ src/main/java/KzeroJun/livestream/auth/dto/response/SignInResponse.java
@@ -1,0 +1,6 @@
+package KzeroJun.livestream.auth.dto.response;
+
+public class SignInResponse {
+
+
+}
