+++ src/main/java/KzeroJun/livestream/auth/service/implement/AuthServiceImpl.java
@@ -1,12 +1,11 @@
-package KzeroJun.livestream.user.service.implement;
+package KzeroJun.livestream.auth.service.implement;
 
-import KzeroJun.livestream.user.dto.request.SignInRequest;
-import KzeroJun.livestream.user.dto.request.SignUpRequest;
+import KzeroJun.livestream.auth.dto.request.SignUpRequest;
 import KzeroJun.livestream.user.entity.User;
 import KzeroJun.livestream.user.exception.DuplicatedEmailException;
 import KzeroJun.livestream.user.exception.DuplicatedNicknameException;
 import KzeroJun.livestream.user.repository.UserRepository;
-import KzeroJun.livestream.user.service.UserService;
+import KzeroJun.livestream.auth.service.AuthService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -15,7 +14,7 @@ import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class UserServiceImpl implements UserService {
+public class AuthServiceImpl implements AuthService {
 
 	private final UserRepository userRepository;
