--- src/main/java/KzeroJun/livestream/user/service/UserService.java
@@ -1,10 +1,0 @@
-package KzeroJun.livestream.user.service;
-
-import KzeroJun.livestream.user.dto.request.SignUpRequest;
-import org.springframework.http.ResponseEntity;
-
-public interface UserService {
-
-	ResponseEntity<Void> signUp(SignUpRequest signUpRequest);
-
-}
